+++ air/src/execution_step/air/never.rs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Fluence Labs Limited
+ * Copyright 2022 Fluence Labs Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,18 @@
  * limitations under the License.
  */
 
-mod ap;
-mod call;
-mod canon;
-mod fail;
-mod fold;
-mod match_;
-mod mismatch;
-mod never;
-mod new;
-mod par;
-mod seq;
-mod xor;
+use super::ExecutionCtx;
+use super::ExecutionResult;
+use super::TraceHandler;
+use crate::log_instruction;
+
+use air_parser::ast::Never;
+
+impl<'i> super::ExecutableInstruction<'i> for Never {
+    fn execute(&self, exec_ctx: &mut ExecutionCtx<'i>, trace_ctx: &mut TraceHandler) -> ExecutionResult<()> {
+        log_instruction!(null, exec_ctx, trace_ctx);
+        exec_ctx.subgraph_complete = false;
+
+        Ok(())
+    }
+}